# libs/@local/graph/api/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "graph",
    "description": "HASH Graph API",
    "contact": {
      "name": "HASH"
    },
    "license": {
      "name": "AGPL-3"
    },
    "version": "0.0.0"
  },
  "paths": {
...
    "/webs": {
      "post": {
        "tags": [
          "Graph",
          "Web"
        ],
        "operationId": "create_web",
        "parameters": [
          {
            "name": "X-Authenticated-User-Actor-Id",
            "in": "header",
            "description": "The ID of the actor which is used to authorize the request",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/UntaggedActorId"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InsertWebIdParams"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "The web was created successfully"
          },
          "500": {
            "description": "Store error occurred"
          }
        }
      }
    },
    "/webs/relationships": {
      "post": {
        "tags": [
          "Graph",
          "Web"
        ],
        "operationId": "modify_web_authorization_relationships",
        "parameters": [
          {
            "name": "X-Authenticated-User-Actor-Id",
            "in": "header",
            "description": "The ID of the actor which is used to authorize the request",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/UntaggedActorId"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/ModifyWebAuthorizationRelationship"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "The relationship was modified for the web"
          },
          "403": {
            "description": "Permission denied"
          }
        }
      }
    },
    "/webs/{web_id}/permissions/{permission}": {
      "get": {
        "tags": [
          "Graph",
          "Web"
        ],
        "operationId": "check_web_permission",
        "parameters": [
          {
            "name": "X-Authenticated-User-Actor-Id",
            "in": "header",
            "description": "The ID of the actor which is used to authorize the request",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/UntaggedActorId"
            }
          },
          {
            "name": "web_id",
            "in": "path",
            "description": "The web ID to check if the actor has the permission",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/EntityId"
            }
          },
          {
            "name": "permission",
            "in": "path",
            "description": "The permission to check for",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/WebPermission"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Information if the actor has the permission for the web",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PermissionResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal error occurred"
          }
        }
      }
    },
    "/webs/{web_id}/relationships": {
      "get": {
        "tags": [
          "Graph",
          "Web"
        ],
        "operationId": "get_web_authorization_relationships",
        "parameters": [
          {
            "name": "X-Authenticated-User-Actor-Id",
            "in": "header",
            "description": "The ID of the actor which is used to authorize the request",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/UntaggedActorId"
            }
          },
          {
            "name": "web_id",
            "in": "path",
            "description": "The web to read the relations for",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/OwnedById"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The relations of the web",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WebRelationAndSubject"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Permission denied"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "EntityId": {
        "type": "string"
      },
...
      "InsertWebIdParams": {
        "type": "object",
        "required": [
          "ownedById",
          "owner"
        ],
        "properties": {
          "ownedById": {
            "$ref": "#/components/schemas/OwnedById"
          },
          "owner": {
            "$ref": "#/components/schemas/WebOwnerSubject"
          }
        },
        "additionalProperties": false
      },
...
      "ModifyRelationshipOperation": {
        "type": "string",
        "description": "Used for mutating a single relationship within the service.",
        "enum": [
          "touch",
          "create",
          "delete"
        ]
      },
      "ModifyWebAuthorizationRelationship": {
        "type": "object",
        "required": [
          "operation",
          "resource",
          "relationAndSubject"
        ],
        "properties": {
          "operation": {
            "$ref": "#/components/schemas/ModifyRelationshipOperation"
          },
          "relationAndSubject": {
            "$ref": "#/components/schemas/WebRelationAndSubject"
          },
          "resource": {
            "$ref": "#/components/schemas/OwnedById"
          }
        }
      },
...
      "OwnedById": {
        "type": "string",
        "format": "uuid"
      },
...
      "PermissionResponse": {
        "type": "object",
        "required": [
          "has_permission"
        ],
        "properties": {
          "has_permission": {
            "type": "boolean"
          }
        }
      },
...
      "UntaggedActorId": {
        "type": "string",
        "format": "uuid"
      },
      "UntaggedTeamId": {
        "type": "string",
        "format": "uuid"
      },
...
      "WebDataTypeViewerSubject": {
        "oneOf": [
          {
            "type": "object",
            "required": [
              "kind"
            ],
            "properties": {
              "kind": {
                "type": "string",
                "enum": [
                  "public"
                ]
              }
            }
          }
        ]
      },
      "WebEntityCreatorSubject": {
        "oneOf": [
          {
            "type": "object",
            "required": [
              "subjectId",
              "kind"
            ],
            "properties": {
              "kind": {
                "type": "string",
                "enum": [
                  "account"
                ]
              },
              "subjectId": {
                "$ref": "#/components/schemas/UntaggedActorId"
              }
            }
          },
          {
            "type": "object",
            "required": [
              "subjectId",
              "kind"
            ],
            "properties": {
              "kind": {
                "type": "string",
                "enum": [
                  "accountGroup"
                ]
              },
              "subjectId": {
                "$ref": "#/components/schemas/UntaggedTeamId"
              }
            }
          }
        ],
        "discriminator": {
          "propertyName": "kind"
        }
      },
      "WebEntityEditorSubject": {
        "oneOf": [
          {
            "type": "object",
            "required": [
              "subjectId",
              "kind"
            ],
            "properties": {
              "kind": {
                "type": "string",
                "enum": [
                  "account"
                ]
              },
              "subjectId": {
                "$ref": "#/components/schemas/UntaggedActorId"
              }
            }
          },
          {
            "type": "object",
            "required": [
              "subjectId",
              "kind"
            ],
            "properties": {
              "kind": {
                "type": "string",
                "enum": [
                  "accountGroup"
                ]
              },
              "subjectId": {
                "$ref": "#/components/schemas/UntaggedTeamId"
              }
            }
          }
        ],
        "discriminator": {
          "propertyName": "kind"
        }
      },
      "WebEntityTypeViewerSubject": {
        "oneOf": [
          {
            "type": "object",
            "required": [
              "kind"
            ],
            "properties": {
              "kind": {
                "type": "string",
                "enum": [
                  "public"
                ]
              }
            }
          }
        ]
      },
      "WebEntityViewerSubject": {
        "oneOf": [
          {
            "type": "object",
            "required": [
              "kind"
            ],
            "properties": {
              "kind": {
                "type": "string",
                "enum": [
                  "public"
                ]
              }
            }
          },
          {
            "type": "object",
            "required": [
              "subjectId",
              "kind"
            ],
            "properties": {
              "kind": {
                "type": "string",
                "enum": [
                  "account"
                ]
              },
              "subjectId": {
                "$ref": "#/components/schemas/UntaggedActorId"
              }
            }
          },
          {
            "type": "object",
            "required": [
              "subjectId",
              "kind"
            ],
            "properties": {
              "kind": {
                "type": "string",
                "enum": [
                  "accountGroup"
                ]
              },
              "subjectId": {
                "$ref": "#/components/schemas/UntaggedTeamId"
              }
            }
          }
        ],
        "discriminator": {
          "propertyName": "kind"
        }
      },
      "WebOwnerSubject": {
        "oneOf": [
          {
            "type": "object",
            "required": [
              "subjectId",
              "kind"
            ],
            "properties": {
              "kind": {
                "type": "string",
                "enum": [
                  "account"
                ]
              },
              "subjectId": {
                "$ref": "#/components/schemas/UntaggedActorId"
              }
            }
          },
          {
            "type": "object",
            "required": [
              "subjectId",
              "kind"
            ],
            "properties": {
              "kind": {
                "type": "string",
                "enum": [
                  "accountGroup"
                ]
              },
              "subjectId": {
                "$ref": "#/components/schemas/UntaggedTeamId"
              }
            }
          }
        ],
        "discriminator": {
          "propertyName": "kind"
        }
      },
      "WebPermission": {
        "type": "string",
        "enum": [
          "change_permission",
          "create_entity",
          "update_entity",
          "view_entity",
          "create_entity_type",
          "create_property_type",
          "create_data_type"
        ]
      },
      "WebPropertyTypeViewerSubject": {
        "oneOf": [
          {
            "type": "object",
            "required": [
              "kind"
            ],
            "properties": {
              "kind": {
                "type": "string",
                "enum": [
                  "public"
                ]
              }
            }
          }
        ]
      },
      "WebRelationAndSubject": {
        "oneOf": [
          {
            "type": "object",
            "required": [
              "subject",
              "relation"
            ],
            "properties": {
              "relation": {
                "type": "string",
                "enum": [
                  "owner"
                ]
              },
              "subject": {
                "$ref": "#/components/schemas/WebOwnerSubject"
              }
            }
          },
          {
            "type": "object",
            "required": [
              "subject",
              "relation"
            ],
            "properties": {
              "relation": {
                "type": "string",
                "enum": [
                  "entityCreator"
                ]
              },
              "subject": {
                "$ref": "#/components/schemas/WebEntityCreatorSubject"
              }
            }
          },
          {
            "type": "object",
            "required": [
              "subject",
              "relation"
            ],
            "properties": {
              "relation": {
                "type": "string",
                "enum": [
                  "entityEditor"
                ]
              },
              "subject": {
                "$ref": "#/components/schemas/WebEntityEditorSubject"
              }
            }
          },
          {
            "type": "object",
            "required": [
              "subject",
              "relation"
            ],
            "properties": {
              "relation": {
                "type": "string",
                "enum": [
                  "entityViewer"
                ]
              },
              "subject": {
                "$ref": "#/components/schemas/WebEntityViewerSubject"
              }
            }
          },
          {
            "type": "object",
            "required": [
              "subject",
              "relation"
            ],
            "properties": {
              "relation": {
                "type": "string",
                "enum": [
                  "entityTypeViewer"
                ]
              },
              "subject": {
                "$ref": "#/components/schemas/WebEntityTypeViewerSubject"
              }
            }
          },
          {
            "type": "object",
            "required": [
              "subject",
              "relation"
            ],
            "properties": {
              "relation": {
                "type": "string",
                "enum": [
                  "propertyTypeViewer"
                ]
              },
              "subject": {
                "$ref": "#/components/schemas/WebPropertyTypeViewerSubject"
              }
            }
          },
          {
            "type": "object",
            "required": [
              "subject",
              "relation"
            ],
            "properties": {
              "relation": {
                "type": "string",
                "enum": [
                  "dataTypeViewer"
                ]
              },
              "subject": {
                "$ref": "#/components/schemas/WebDataTypeViewerSubject"
              }
            }
          }
        ],
        "discriminator": {
          "propertyName": "relation"
        }
      }
    }
  },
...
}